+++ src/test_typing_extensions/ext.py
@@ -1,0 +1,2 @@
+# We just import the ext module to check for syntax or import problems.
+from typing_extensions.ext import *